[US_ND] Ingest StateProgramAssignment (Recidiviz/recidiviz-data#28356)
## Description of the change

Ingest information from the Elite system tables
elite_OffenderProgramProfiles into state_program_assignment. It is
important to note that some of these "program assignments" are not
programs at all, but case manager assignments, or work assignments. All
expected programs are present (T4C, CBISA, etc), but there is a wide
array of program descriptions and the word "program" itself seems to be
used as a bit of a catch-all here. There is no easy way to filter out
the irrelevant assignments, so they are all ingested here.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#27856
## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7d6d53e356956efc4f297668f8a587ed86feced9