[IID Cleanup] STATE schema test cleanup (Recidiviz/recidiviz-data#28396)
## Description of the change

When we update `SchemaType.has_cloud_sql_instance()` to return false for
the `STATE` schema, all tests that actually reference a postgres
instance of the `STATE` schema will fail.

This PR cleans up / removes unnecessary tests that will fail when that
flag is flipped. Most are irrelevant because they check things like
constraints that will never be enforced via Postgres. I added one new
test that crashes if you defined the state schema in a way that does not
compile (e.g. a relationship field that references an entity that
doesn't exist).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#20930.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: c8a8e9918a8e4ebe20954cf19041e23df79e637d