[Monitoring] Bump INCIDENT_START_DATE_LOOKBACK to 21 (Recidiviz/recidiviz-data#28416)
## Description of the change

In Recidiviz/recidiviz-data#22895 we lowered the lookback threshold for open incidents from 30
days to 7 days to alleviate some scale issues we were seeing with the
monitoring DAG. However, this means that issues don't auto-resolve when
they are open for longer than 7 days. We recently had some
state-specific DAG failures that went on for over a month and each week
a new set of incidents would be re-opened for the failing nodes.

Since Recidiviz/recidiviz-data#22895 made some other performance improvements outside of
lowering the lookback threshold, we're going to increase this to 21 days
to make auto-resolving better and keep an eye on the monitoring DAG to
make sure it can handle the load.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

See this thread:
https://recidiviz.slack.com/archives/C03JV24TBDF/p1710857845588049

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 274455c5a291972609c71fcfecc55edab8234599