[Justice Counts] Clean up is_metric_disabled. (Recidiviz/recidiviz-data#28367)
## Description of the change

It turns out that is_metric_disabled() is only called once in the
codebase. In this one invocation, dimension_id=None. This means that we
can safely remove any logic pertaining to the dimension_id != None case.

NOTE that this means that is_metric_disabled is now really just checking
whether the overall metric is disabled. Let's change the name to make
that behavior clear.

Making this change in a separate PR (as opposed to within a refactoring
PR) so that it can be reviewed more carefully.

## Testing

Tested this locally and ran all unit tests. But also just walked through
the logic and made sure that the behavior is the same in the
dimension_id == None case.
GitOrigin-RevId: 7cd83036f1c9044538049e2683e946b227ac1be8