[Dataflow] Use person demographics view for race/ethnicity instead of output from pipeline (Recidiviz/recidiviz-data#28415)
## Description of the change

This is the prework for tearing out the demographics calculation from
the dataflow metric pipeline code. This changes the most recent views to
pull the race/ethnicity from the `person_demographics` view.

We are focusing just on race/ethnicity for now, as that is currently the
expensive part of the dataflow pipeline. The way that we have structured
the reading of state population resident counts causes a lot of data to
be shuffled, and that is only used for race/ethnicity.

Sandbox output lives in `colin_dataflow_metrics_materialized`. I only
checked a handful of metrics, but for those it showed no differences.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Start of Recidiviz/recidiviz-data#25565

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: b567a0575f49c80ba4b70429f07f53be00ddd60a