[Justice Counts] Refactor SpreadsheetInterface to use MetricSettingInterface. (Recidiviz/recidiviz-data#28376)
## Description of the change

Refactor SpreadsheetInterface so that it relies on calls to
MetricSettingInterface instead of calls to DatapointInterface.

Try to review PR Recidiviz/recidiviz-data#28367 first, which is a smaller change that is
included in this PR. I have it in a separate PR so that the change can
receive closer inspection.

Zenhub Ticket Recidiviz/recidiviz-data#27956

## Testing

Tested locally and on playtesting for bulk upload. Also tested on
Superagencies with child agencies to make sure that functionality stays
the same.

Modified unit tests and everything works as expected.

Here's the playtesting URL with this PR for testing:

https://test---publisher-web-b47yvyxs3q-uc.a.run.app/

GitOrigin-RevId: 3814f7f1b3c506a5c8781f35ec41d2fac29b8f89