[US_IX] Ingest supervision violations from ind_Offender_QuestionnaireTemplate (Recidiviz/recidiviz-data#28077)
## Description of the change

For Idaho, we already ingest supervision violations via the
scl_Violation* tables via the ingest view
[view_supervision_violation](https://github.com/Recidiviz/recidiviz-data/blob/e3a875ac57be5915a4c990b29ec29f87b2fdf02a/recidiviz/ingest/direct/regions/us_ix/ingest_views/view_supervision_violation.py),
but we found that officers did not seem to be entering many violations
in those tables. In addition, of the violations they entered, violation
type was often missing. Based on conversations with Idaho, it seems like
there may be additional supervision violation information saved in the
violation survey of the questionnaire section of Atlas.

Thus, we decided to ingest supervision violations from that area of
Atlas as well, and see how much additional information about violations
that provides.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#27502

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f10421158d497c79a424367404d8c63cad4a8151