[Insights] Add absconders label to Configuration DB entity and frontend (Recidiviz/recidiviz-data#28424)
## Description of the change

Adds the absconder label to the Configuration entity in the DB and
generates the corresponding DB migration, which includes using a default
value to hydrate this column for existing DB entities. Also adds this
column where necessary in the admin panel BE and FE and documents
necessary steps in `outliers/schema.py`. I tested this by running the
`load_local_db` script for the `configurations` table with the fixture
data in `configurations.csv` without the absconders_label field and
ensuring that each row had the default value applied

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#27981

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f31c9c2483d587c07cfb29ae7dc601c08d59063f