[US_AZ] Ingest Transition Release eligibility & due dates  (Recidiviz/recidiviz-data#28431)
## Description of the change

Anyone who has been approved or tentatively approved for the Transition
Release Program will have a related `StateTaskDeadline`. The
`eligibility_date` is the date they became eligible for transition
release. The `due_date` is the date they were released on transition
release, if that was in the past, or the date they become eligible for
transition release if they are currently incarcerated. If they were
released before they were noted as eligible, then their `eligible_date`
is updated to be the same as their `due_date`, based on the assumption
that they were in fact eligible at the time they were released.

Tests:

- [x] ingest view is deterministic
- [x] [sandbox ingest
pipeline](https://console.cloud.google.com/dataflow/jobs/us-west1/2024-03-26_10_20_36-12673211254911709124;graphView=0?project=recidiviz-staging)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 5dc9dd36d399b6011e0bb525461737a17ff15257