[US_AZ] Ingest StateSentenceGroup (Recidiviz/recidiviz-data#28427)
## Description of the change

This entity holds information about the controlling offense for a given
commitment to incarceration. There can be many offenses tied to a single
commitment, but one of them controls the sentence as a whole. This
offense is the one that determines the release dates. The sentence group
external ID is the ADCRR system-generated ID for the _commitment_, along
with the sequence number of this particular set of release dates for
that commitment. The commitment ID alone can also be found on
`state_sentence.sentence_group_external_id`, where the same
`sentence_group_external_id` can be attached to multiple
`state_sentence` entities when there are multiple offenses per
commitment.

The mappings are:

- `parole_eligibility_date_external`: CommunitySupervisionBeginDate, the
earliest date a person can be released to community supervision outside
of a transition release.
- `projected_full_term_release_date_min_external`:
EarnedReleaseCreditDate
- `projected_full_term_release_date_max_external`: If a person is not
eligible for community supervision, then the latest they can be released
from prison is their Sentence Expiration Date (SED). If they _are_
eligible for community supervision, then the latest they can be released
from prison is their Community Supervision End Date (CSED), which is
115% of the SED. If a CSED exists for the controlling offense, then it
is stored here. If no CSED exists, the SED is stored here.

tests:
- [x] ingest view is deterministic
- [x] [sandbox ingest
pipeline](https://console.cloud.google.com/dataflow/jobs/us-west1/2024-03-26_08_38_18-1317953425761724098;graphView=1?project=recidiviz-staging)
## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: c7992575233bc8f374b991900f7b95a2bc2874ef