[Workflows][US_PA] Enable workflows exports for PA (Recidiviz/recidiviz-data#28428)
## Description of the change

This PR adds US_PA to the staff, supervision staff, and client records.
It also enables exports of PA workflows views.

The staff query is based off of the mostly-state-agnostic US_IX staff
query. To my eyes, this one is completely state agnostic, so that's a
good sign for our ultimate goals. (@danawillow)

This PR was tested by running the queries by hand and saving the results
in `kesich_scratch`. See `clients_pa_only` and `staff_pa`. These views,
along with the opportunity query, were loaded into staging firestore
using `recidiviz.tools.workflows.run_etl_from_local_branch`. I then
loaded up the staging dash and saw that officers are loaded into the
search bar in PA and they have clients that appear correctly in the "all
clients" views. No clients currently show up for the opportunity, but
the console shows that's due to a mismatch between zod and firestore
which can be addressed in a separate PR, probably on the front-end.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#28312

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] ~Tests have been written to cover the code changed/added as part
of this pull request~ N/A (yet) BQ change

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: a51310b3e3a5504f0d2ad2ffeeefadeb65470d06