[US_MI][Workflows] Updates to reclassification to general opportunity (Recidiviz/recidiviz-data#28221)
## Description of the change

Based on trusted tester feedback and failing validations, we made the
following opportunity and criteria changes:

**opportunity changes:**
_We combined the overdue in disciplinary and overdue in temporary
criteria to contribute to one combined criteria since residents often
transition between this compartments and it makes sense for these both
to live under this umbrella opportunity_
1. Combined the `detention_sanction_timeframe_has_expired` logic with
the `in_temporary_solitary_confinement_at_least_30_days` logic so both
of those criteria contribute to the
`eligible_for_reclassification_from_solitary_to_general` criteria and
the **reclass opp** (we deleted the temporary solitary confinement
opportunity)
2. Updated the completion event so that transitions out of solitary from
both `TEMPORARY` and `DISCIPLINARY` housing unit types are captured.

**updates to the detention criteria**
- Created sub sessions where stays in
`DISCIPLINARY_SOLITARY_CONFINEMENT` are associated with detention
sanction spans and for each sub session, the MAX `critical_date` is used
to account for overlapping detention sanctions.
- This results in residents being immediately eligible for reclass while
in detention if there is no detention sanction placed, and allows for
back to back sanctions which will expire when the detention sanction
expires (vs. x days from the `DISCIPLINARY_SOLITARY_CONFINEMENT` start).
- This has pros and cons since if a residents is in detention for x
amount of days before the detention sanction is in place, those days
won't count towards the 10 or 20 days limited by policy. However, this
allows us to effectively associate overlapping detention sanction spans
within a solitary stay and have them expire in a logical way.
Additionally, since new sanctions can be made while a resident is in
solitary, this logic is in line with feedback from TT.

**outstanding questions**
- I noticed that if a detention sanction timeframe expires and the
resident gets transferred to a different type of solitary confinement
(ie `TEMPORARY_SOLITARY_CONFINEMENT`), then the resident will not be
eligible for reclass until they have been in temporary for > 30 days,
but I wonder if they should be immediately eligible.
https://github.com/Recidiviz/recidiviz-data/issues/28219

> Update: after emailing Jeremy (head of facilities) it seems like
detention --> temporary restarts the clock, so this logic is in line
with their policy.

Tested in `shuff` and examples of criteria logic
[here](https://docs.google.com/spreadsheets/d/1d-kz1t6JvjAlLPHAX1WtKDuRjSzVdAtGWZGGMPGQ5G0/edit#gid=1475427862)
## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#27710
Closes Recidiviz/recidiviz-data#27872
Closes Recidiviz/recidiviz-data#28204

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7ada66dbd7a22ca1a638aa9e8b110ad201a47d68