[US_MI] Set always_historical_export to True for COMS case managers table (Recidiviz/recidiviz-data#28473)
## Description of the change

We've discovered some records being deleted out of this table on MI's
side. Since we are still getting full historicals of this table for now,
let's set always_historical_export to True so that we only consider data
from the latest transfer in the latest view.

NOTE: At this point, MI should be revoking the permissions that was
allowing agents to delete records, and so hopefully(?) we shouldn't see
this issue in the future? But at some point, the COMS developers might
also switch this transfer to being two week diffs instead of full
historicals, so if that were the case, we'd have to prune out the old
historical copies and keep just the latest, and then set
`always_historical_export` back to False. Still discussing with MI the
implications of all this.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e1eccaa3290bf22d7b6fb52c9425ab62d07db287